New AddressInfoGrid Note the experimentation with subcomponents
